Download Links
Update the Load Links to download and cache the data.
By default, the article data will be stored at "./GDELT_REPOSITORY/articles"
The filename is generated by a (relatively) stable function using only
the URL so that we do not need to download unecessary articles.

A good idea may be to use a hash function of the link data to generate
the file names instead.